backend/requirements/base.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
django-cors-headers
whitenoise
Pillow
boto3
django-storages
celery
redis
django-celery-beat